+++ __init__.py
@@ -1,0 +1,1 @@
+"""AstrBot AIRSS plugin package."""